refactor(dsh-mcp-manager): 阶段6批次1——types/ 域收敛（server/status/ui/middleware-types + interface.ts，Part of #664）
src/types.ts 拆 server.ts/status.ts/ui.ts，middleware-types.ts 迁入 types/，
新增 types/interface.ts 门面（D10）；19 个引用文件全切
types/interface.ts，根 types.ts 删除。纯类型搬移，零行为变更（type-only）。

Related issue #664
