+++ packages/dsh-mcp-manager/src/catalog/cache-view.ts
@@ -12,7 +12,7 @@ import { existsSync, statSync } from "node:fs";
 import { join } from "node:path";
 import { dshHome } from "../../../../shared/dsh-home.js";
 import type { McpMiddleware } from "../middleware.ts";
-import type { ServerConfig } from "../types.ts";
+import type { ServerConfig } from "../types/interface.ts";
 import { readCatalogServerFromDisk } from "../middleware-state.ts";
 import { normalizedProjectRoot, SCOPE_PROJECT, MIDDLEWARE_GLOBAL_ROOT } from "../workspace/interface.ts";
 import { summarizeToolDescriptions } from "./entries.ts";
