+++ packages/dsh-mcp-manager/src/config-schema.ts
@@ -9,7 +9,7 @@ import z from "schemastery";
 import { clampZIndexBase, DEFAULT_Z_INDEX_BASE } from "./placement-math.ts";
 import { DEFAULT_ANNOUNCE_CATALOG, DEFAULT_CATALOG_MAX_ENTRIES } from "./catalog/interface.ts";
 import { DEFAULT_RESULT_TRUNCATE_BYTES } from "./supervisor.ts";
-import type { ClientUiConfig, UiPlacementConfig } from "./types.ts";
+import type { ClientUiConfig, UiPlacementConfig } from "./types/interface.ts";
 
 /** 空 description 工具的条件拼接默认开启。 */
 export const DEFAULT_ENHANCE_EMPTY_DESCRIPTIONS = true;
