+++ packages/dsh-mcp-manager/src/import.ts
@@ -7,7 +7,7 @@
  * 由 lib/index.js 组合根 re-export。
  */
 
-import type { ServerConfig } from "./types.ts";
+import type { ServerConfig } from "./types/interface.ts";
 
 // ------------------------------------------- mcpServers JSON 条目映射
