+++ packages/dsh-mcp-manager/src/pipeline/redact.ts
@@ -6,7 +6,7 @@
  * 错误消息中出现的是原始 URL 字节串（percent-encoding 形态），URL parse 拿到
  * decoded 形态，两者都须命中（否则 percent-encoding 绕过回归，评审 B8 修正）。
  */
-import type { ServerConfig } from "../types.ts";
+import type { ServerConfig } from "../types/interface.ts";
 
 /** 注册单个敏感值：decoded 形态 + percent-encoded raw 形态双注册。 */
 function addSecretPair(secrets: Set<string>, value: string): void {
