+++ packages/dsh-mcp-manager/src/transport.ts
@@ -14,7 +14,7 @@
 
 import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
 import { StreamableHTTPClientTransport } from "@modelcontextprotocol/sdk/client/streamableHttp.js";
-import type { ServerConfig } from "./types.ts";
+import type { ServerConfig } from "./types/interface.ts";
 
 // ------------------------------------------------- env 展开 / 子进程环境
